Further restructuring and consolidate constants
