+++ corebehrt/functional/trainer/collate.py
