+++ corebehrt/main/pretrain.py
@@ -6,12 +6,13 @@ from corebehrt.functional.features.split import (
     load_train_val_split,
     split_pids_into_train_val,
 )
-from corebehrt.functional.load import load_vocabulary
-from corebehrt.functional.save import save_pids_splits
+from corebehrt.functional.io_operations.load import load_vocabulary
+from corebehrt.functional.io_operations.save import save_pids_splits
 from corebehrt.functional.setup.args import get_args
 from corebehrt.functional.setup.model import load_model_cfg_from_checkpoint
 from corebehrt.functional.trainer.setup import replace_steps_with_epochs
 from corebehrt.modules.preparation.dataset import MLMDataset
+
 from corebehrt.modules.preparation.prepare_data import DatasetPreparer
 from corebehrt.modules.setup.config import load_config
 from corebehrt.modules.setup.directory import DirectoryPreparer
