+++ corebehrt/modules/preparation/prepare_data.py
@@ -10,14 +10,14 @@ import torch
 from dask.diagnostics import ProgressBar
 
 from corebehrt.functional.cohort_handling.outcomes import get_binary_outcomes
-from corebehrt.functional.load import load_vocabulary
+from corebehrt.functional.io_operations.load import load_vocabulary
 from corebehrt.functional.preparation.convert import dataframe_to_patient_list
 from corebehrt.functional.preparation.filter import (
     censor_patient,
     exclude_short_sequences,
 )
 from corebehrt.functional.preparation.truncate import truncate_patient
-from corebehrt.functional.save import save_vocabulary
+from corebehrt.functional.io_operations.save import save_vocabulary
 from corebehrt.functional.utils import (
     get_abspos_from_origin_point,
     get_background_length,
@@ -27,13 +27,14 @@ from corebehrt.functional.utils import (
 from corebehrt.modules.cohort_handling.patient_filter import filter_df_by_pids
 from corebehrt.modules.preparation.dataset import PatientDataset
 from corebehrt.modules.setup.config import Config, load_config
-from corebehrt.modules.setup.constants import ABSPOS_COL, PID_COL, TIMESTAMP_COL
+from corebehrt.functional.constants import ABSPOS_COL, PID_COL, TIMESTAMP_COL
 from corebehrt.modules.setup.directory import (
     INDEX_DATES_FILE,
     OUTCOMES_FILE,
     PID_FILE,
     PROCESSED_DATA_DIR,
 )
+
 
 logger = logging.getLogger(__name__)  # Get the logger for this module
