--- corebehrt/modules/setup/constants.py
@@ -1,6 +1,0 @@
-### Columns ###
-PID_COL = "PID"
-TIMESTAMP_COL = "TIMESTAMP"
-BIRTHDATE_COL = "BIRTHDATE"
-DEATHDATE_COL = "DEATHDATE"
-ABSPOS_COL = "abspos"
