+++ corebehrt/modules/trainer/trainer.py
@@ -5,7 +5,7 @@ from collections import namedtuple
 
 from torch.utils.data import DataLoader, Dataset
 from corebehrt.modules.setup.config import Config, instantiate_class
-from corebehrt.functional.collate import dynamic_padding
+from corebehrt.functional.trainer.collate import dynamic_padding
 from corebehrt.modules.monitoring.metric_aggregation import (
     compute_avg_metrics,
     save_curves,
